Make flux9s lib compatible
