+++ Cargo.toml
@@ -62,6 +62,9 @@ exclude = [
 ]
 
 [features]
+default = ["tui"]
+# Terminal UI (ratatui + crossterm) — disable for headless library usage
+tui = ["dep:ratatui", "dep:crossterm"]
 # Vendor OpenSSL for static linking (needed for musl targets)
 vendored-openssl = ["openssl/vendored"]
 
@@ -84,9 +87,9 @@ k8s-openapi = { version = "0.26.0", default-features = false, features = [
 tokio = { version = "1.48", features = ["full"] }
 futures = "0.3"
 
-# TUI framework
-ratatui = "0.29"
-crossterm = "0.29"
+# TUI framework (optional — enable with "tui" feature)
+ratatui = { version = "0.29", optional = true }
+crossterm = { version = "0.29", optional = true }
 
 # Error handling
 anyhow = "1.0"
@@ -161,3 +164,13 @@ path = "tests/field_extraction.rs"
 [[test]]
 name = "trace_tests"
 path = "tests/trace_tests.rs"
+
+[[test]]
+name = "snapshot_tests"
+path = "tests/snapshot_tests.rs"
+required-features = ["tui"]
+
+[[test]]
+name = "navigation_tests"
+path = "tests/navigation_tests.rs"
+required-features = ["tui"]
